webapp: Fix three bugs
Project-from-specimen flow incorrectly handled the case where the user
has more than one project linked to the specimen but all of them are
identical to the specimen.

When creating a project from a zipfile, the assets were not added in a
deterministic order.  For "flat" programs this isn't terrible, but for
"per-method" programs it matters.

Changing the order of (say) costumes in a per-method program did not
update that project's mtime.